+++ HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ~~~~~~~~~~
 
+- Group images by dates.
+  [msom]
+
 0.3.0 (2015-08-03)
 ~~~~~~~~~~~~~~~~~~~